Move global scrollbar file
